Moved some data extraction code out of ViewController.swift to its own file, for clarity
